converted it to mvc architecture
