+++ package.json
@@ -12,11 +12,13 @@
   "author": "Google LLC",
   "license": "Apache-2.0",
   "dependencies": {
+    "@types/dotenv": "^8.2.3",
     "@types/express": "^5.0.0",
     "@types/jimp": "^0.2.28",
     "@types/multer": "^1.4.12",
     "@types/qrcode": "^1.5.5",
     "@types/xlsx": "^0.0.36",
+    "dotenv": "^16.4.7",
     "express": "^4.19.2",
     "jimp": "^0.22.12",
     "multer": "^1.4.5-lts.1",
