+++ src/workers/qr-generation-worker.ts
@@ -1,6 +1,6 @@
 import { parentPort, workerData } from "worker_threads";
-import { QRCodeGenerator } from "./qrgenerator.service";
-import { logger } from "./logger-service";
+import { QRCodeGenerator } from "../services/qrgenerator.service";
+import { logger } from "../services/logger.service";
 
 async function generateQRCodes() {
   const { filePath, urlTag, nameTag, logoPath, jobId } = workerData;
